% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateAverageCurrentFlowBetweeness}
\alias{calculateAverageCurrentFlowBetweeness}
\title{calculate the average current flow betweeness}
\usage{
calculateAverageCurrentFlowBetweeness(graphLaplacianInv, adjacency)
}
\arguments{
\item{graphLaplacianInv}{The psuedoinverse of the graph laplacian. The graph MUST be fully connected - I make no checks for you ...}

\item{adjacency}{The adjacency matrix of a (undirected) graph}
}
\value{
currentFlowBetweeness A vector of current flow betweeness, averaged over all source-sink pairs
}
\description{
We take advantage of the notion of resistor network to define current-flow. It is defined as
\deqn{F_{i}^{(st)}=\frac{1}{2} \sum_{j} A_{ij} \left | V_{i}^{(st)} - V_{j}^{(st)}  \right |}
Given nodes \eqn{i} and \eqn{j} and a unit of current injected from source \eqn{s} and removed from target \eqn{t}.
The current flow betweeness is defined as: \deqn{b_{i} = \frac{\sum_{s<t} F_{i}^{(st)} }{\frac{1}{2}n(n-1))}}. Hence
averageing over all \eqn{s} - \eqn{t} pairs.
}
\references{
\url{http://www.sci.unich.it/~francesc/teaching/network/flowcentrality.html}
}
