% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateResistiveDistances}
\alias{calculateResistiveDistances}
\title{generate the pairwise resistive distance}
\usage{
calculateResistiveDistances(graphLaplacianInv)
}
\arguments{
\item{graphLaplacianInv}{The psuedoinverse of the graph laplacian. The graph MUST be fully connected - I make no checks for you ...}
}
\value{
currentFlowDistances A matrix of pairwise resistive distances from the network
}
\description{
We take advantage of the notion of resistor network to define current-flow distance.
Given nodes \eqn{i} and \eqn{j}, the resistance distance \eqn{R_{i,j}}{R_ij} between \eqn{i} and \eqn{j}
is the effective (or equivalent) resistance between \eqn{i} and \eqn{j} when a unit of current is injected
from source \eqn{s} and removed from target \eqn{t}.
}
\references{
\url{http://www.sci.unich.it/~francesc/teaching/network/flowcentrality.html}
}
