% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentFlowCentrality.R
\docType{package}
\name{currentFlowCentrality}
\alias{currentFlowCentrality}
\alias{currentFlowCentrality-package}
\title{currentFlowCentrality: A package for a betweeness cetrality based on current flow}
\description{
Most centralities are based on the idea of counting shortest paths from A to B and then iterating over all AB pairs.
Newman et al suggests the idea of using the principle of electrical current flow, hence counting all paths.
}
\references{
Newman, M. E. J. (2005). A measure of betweenness centrality based on random walks. Social Networks, 27(1)
}
\author{
Adam Sardar
}
