% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcPinv.R
\docType{methods}
\name{dcPinv}
\alias{dcPinv}
\alias{dcPinv,matrix-method}
\alias{dcPinv,Matrix-method}
\title{Calculate the Moore-Penrose psuedo-inverse of a matrix using the divide & conquer algorithm in LAPACK}
\usage{
dcPinv(matrixIn)

\S4method{dcPinv}{matrix}(matrixIn)

\S4method{dcPinv}{Matrix}(matrixIn)
}
\arguments{
\item{matrixIn}{Matrix for which to calculate the psuedo-inverse}
}
\value{
ginv The matrix representation of the psuedo-inverse
}
\description{
The ubiquitous LAPACK library provides several implementations for the singular-value decomposition (SVD), which underlies
psuedo-inverse calculation. The divide and conquer algorithm is particularly quick.
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: Psuedo-inverse of base matrix class objects

\item \code{Matrix}: Psuedo-inverse of Matrix class objects
}}

\references{
\url{http://gallery.rcpp.org/articles/divide-and-concquer-svd/}

\url{https://en.wikipedia.org/wiki/Moore-Penrose_pseudoinverse}
}
