% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dcPinvDense}
\alias{dcPinvDense}
\title{generate the Moore-Penrose psueoinverse using the armadillo package}
\usage{
dcPinvDense(X)
}
\arguments{
\item{X}{A matrix to be psuedo-inverted}
}
\value{
An armadillo (dense) matrix containg the psuedo-inverse values
}
\description{
The Armadillo package provides a low level interface to native LAPACK/BLAS routines
and we plan to use the divide-and-conquer implementation of SVD to generate the psuedo-inverse
}
\references{
\url{http://en.wikipedia.org/wiki/Moore-Penrose_pseudoinverse}

\url{http://gallery.rcpp.org/articles/divide-and-concquer-svd/}
}
