% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{singleSourceTargetFlow}
\alias{singleSourceTargetFlow}
\title{for a single source and single sink node of unit current, calculate the per-node flow}
\usage{
singleSourceTargetFlow(Gpinv, adjacency, sourceIndex, sinkIndex)
}
\arguments{
\item{Gpinv}{The psuedoinverse of the (undirected) graph laplacian. The graph MUST be fully connected - I make no checks for you ...}

\item{adjacency}{The adjacency matrix of a (undirected) graph}

\item{sourceIndex}{The integer index of the source of unit flow}

\item{sinkIndex}{The integer index of the sink of unit flow}
}
\value{
currentFlowDistances A column matrix of resistive distances from the network
}
\description{
We take advantage of the notion of resistor network to define current-flow. It is defined as
\deqn{F_{i}^{(st)}=\frac{1}{2} \sum_{j} A_{ij} \left | V_{i}^{(st)} - V_{j}^{(st)}  \right |}
Given nodes \eqn{i} and \eqn{j} and a unit of current injected from source \eqn{s} and removed from target \eqn{t}.
}
\references{
\url{http://www.sci.unich.it/~francesc/teaching/network/flowcentrality.html}
}
