% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplInvApprox.R
\name{spNorm}
\alias{spNorm}
\title{Calculates the 2-norm or spectral norm of a matrix}
\usage{
spNorm(mat)
}
\arguments{
\item{mat}{Input matrix}
}
\value{
norm The spectral norm of the matrix
}
\description{
The spectral norm of a matrix is the largest singular value
}
