% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{standardPinv}
\alias{standardPinv}
\title{generate the Moore-Penrose psueoinverse using the armadillo package}
\usage{
standardPinv(X)
}
\arguments{
\item{X}{A matrix to be psuedo-inverted}
}
\value{
An armadillo (dense) matrix containg the psuedo-inverse values
}
\description{
The Armadillo package provides a low level interface to native LAPACK/BLAS routines
It provides several implementations of SVD and this function uses the standard algorithm
}
\references{
\url{http://en.wikipedia.org/wiki/Moore-Penrose_pseudoinverse}

\url{http://gallery.rcpp.org/articles/divide-and-concquer-svd/}
}
